requirements.txt
azure-functions
twilio
requests
openai